requirements.txt
aio-pika==9.3.1
loguru==0.7.2
pytest==7.4.3
pytest-asyncio==0.23.2
python-dotenv==1.0.0
roslz4
--extra-index-url https://rospypi.github.io/simple/

pydantic~=2.5.3